+++ docs/html/search/functions_2.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['eeprom_5fwrite_5fbyte_0',['eeprom_write_byte',['../group__eeprom.html#ga80807aaa414692b04963d0ef4d5f5d0b',1,'eeprom_write_byte(uint8_t addr_high, uint8_t addr_low, uint8_t value, uint8_t *error_ptr):&#160;eeprom.c'],['../group__eeprom.html#ga80807aaa414692b04963d0ef4d5f5d0b',1,'eeprom_write_byte(uint8_t addr_high, uint8_t addr_low, uint8_t value, uint8_t *error_ptr):&#160;eeprom.c']]]
+  ['getchar_0',['getchar',['../group__uart.html#ga3e29caa20f7cffe18f410f01278905a8',1,'getchar(void):&#160;uart.c'],['../group__uart.html#ga3e29caa20f7cffe18f410f01278905a8',1,'getchar(void):&#160;uart.c']]]
 ];